% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_hostname}
\alias{build_hostname}
\title{Build host name/database}
\usage{
build_hostname(db)
}
\arguments{
\item{db}{A \code{character} string. Possible values are 'string' (default)
or 'stitch'.}
}
\value{
A \code{character} string.
}
\description{
Build host name/database
}
\examples{
# build hostname for string
build_hostname('string')

}
