% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{enrichment}
\alias{enrichment}
\title{Getting functional enrichment}
\source{
https://string-db.org/cgi/help.pl?subpage=api%23getting-functional-enrichment
}
\usage{
enrichment(identifiers = NULL, background_string_identifiers,
  species = 9606, caller_identity)
}
\arguments{
\item{identifiers}{A \code{character} string.}

\item{background_string_identifiers}{A \code{character} string.}

\item{species}{A \code{numeric}.}

\item{caller_identity}{A \code{character} string.}
}
\value{
A \code{tibble}.
\describe{
 \item{category}{term category (e.g. GO Process, KEGG pathways)}
 \item{term}{enriched term (GO term, domain or pathway)}
 \item{number_of_genes}{number of genes in your input list with the term
 assigned}
 \item{number_of_genes_in_background}{total number of genes in the
 background proteome with the term assigned}
 \item{ncbiTaxonId}{NCBI taxon identifier}
 \item{inputGenes}{gene names from your input}
 \item{preferredNames}{common protein names (in the same order as your
 input Genes)}
 \item{p_value}{raw p-value}
 \item{fdr}{False Discovery Rate}
 \item{description}{description of the enriched term}
}
}
\description{
Performs the enrichment analysis of your set of proteins for the Gene
Ontology, KEGG pathways, UniProt Keywords, PubMed publications, Pfam,
InterPro and SMART domains.
}
\examples{
\dontrun{
# make an enrichment request
enrichment(identifiers = c('trpA','trpB','trpC','trpE','trpGD'))
}

}
\seealso{
\code{\link{get_string_ids}}
\code{\link{network}}
\code{\link{interaction_partners}}
\code{\link{homology}}
\code{\link{homology_best}}
\code{\link{functional_annotation}}
\code{\link{ppi_enrichment}}

Other API methods: \code{\link{functional_annotation}},
  \code{\link{get_string_ids}},
  \code{\link{homology_best}}, \code{\link{homology}},
  \code{\link{interaction_partners}},
  \code{\link{network}}, \code{\link{ppi_enrichment}}
}
\concept{API methods}
