% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{functional_annotation}
\alias{functional_annotation}
\title{Retrieving functional annotation}
\source{
https://string-db.org/cgi/help.pl?subpage=api%23retrieving-functional-annotation
}
\usage{
functional_annotation(identifiers = NULL, species = 9606,
  allow_pubmed = 0, caller_identity)
}
\arguments{
\item{identifiers}{A \code{character} string.}

\item{species}{A \code{numeric}.}

\item{allow_pubmed}{A \code{logical} in the form '1' (default) or '0'.}

\item{caller_identity}{A \code{character} string.}
}
\value{
A \code{tibble}.
\describe{
 \item{category}{term category (e.g. GO Process, KEGG pathways)}
 \item{term}{enriched term (GO term, domain or pathway)}
 \item{number_of_genes}{number of genes in your input list with the term
 assigned}
 \item{ratio_in_set}{ratio of the proteins in your input list with the
 term assigned}
 \item{ncbiTaxonId}{NCBI taxon identifier}
 \item{inputGenes}{gene names from your input}
 \item{preferredNames}{common protein names (in the same order as your
 input Genes)}
 \item{description}{description of the enriched term}
}
}
\description{
Gets the functional annotation (Gene Ontology, UniProt Keywords, PFAM,
INTERPRO and SMART domains) of your list of proteins.
}
\examples{
\dontrun{
# make a functional_annotation request
functional_annotation(identifiers = 'cdk1')
}

}
\seealso{
\code{\link{get_string_ids}}
\code{\link{network}}
\code{\link{interaction_partners}}
\code{\link{homology}}
\code{\link{homology_best}}
\code{\link{enrichment}}
\code{\link{ppi_enrichment}}

Other API methods: \code{\link{enrichment}},
  \code{\link{get_string_ids}},
  \code{\link{homology_best}}, \code{\link{homology}},
  \code{\link{interaction_partners}},
  \code{\link{network}}, \code{\link{ppi_enrichment}}
}
\concept{API methods}
