% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_abstracts}
\alias{get_abstracts}
\title{Get abstracts request}
\usage{
get_abstracts(identifier = NULL, limit = 5, format = "pmid",
  db = "string")
}
\arguments{
\item{identifier}{A \code{character} vector.}

\item{limit}{A \code{numeric}.}

\item{format}{A \code{character} string. Possible values are 'pmid'
(default) or 'colon'.}

\item{db}{A \code{character} string. Possible values are 'string' (default)
or 'stitch'.}
}
\value{
A \code{tibble}.
}
\description{
Get abstracts request
}
\examples{
\dontrun{
# make abstracts request
get_abstracts(identifier = c('4932.YML115C', '4932.YJR075W', '4932.YEL036C'))
}

}
