% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_actions}
\alias{get_actions}
\title{Get actions request}
\usage{
get_actions(identifier = NULL, limit = 5, required_score,
  additional_network_nodes, db = "string")
}
\arguments{
\item{identifier}{A \code{character} vector.}

\item{limit}{A \code{numeric}.}

\item{required_score}{A \code{numeric}.}

\item{additional_network_nodes}{A \code{numeric}}

\item{db}{A \code{character} string. Possible values are 'string' (default)
or 'stitch'.}
}
\value{
A \code{tibble}.
}
\description{
Get actions request
}
\examples{
\dontrun{
# make actions request
get_actions(identifier = 'ADD')
}

}
