% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_interactions}
\alias{get_interactions}
\title{Get interactions request}
\usage{
get_interactions(identifier = NULL, limit = 5, required_score,
  additional_network_nodes, db = "string")
}
\arguments{
\item{identifier}{A \code{character} vector.}

\item{limit}{A \code{numeric}.}

\item{required_score}{A \code{numeric}.}

\item{additional_network_nodes}{A \code{numeric}}

\item{db}{A \code{character} string. Possible values are 'string' (default)
or 'stitch'.}
}
\value{
A \code{tibble}
}
\description{
Get interactions request
}
\examples{
\dontrun{
# make interactions request
get_interactions(identifier = 'ADD')
}

}
