% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_string_ids}
\alias{get_string_ids}
\title{Mapping identifiers}
\source{
https://string-db.org/cgi/help.pl?subpage=api%23mapping-identifiers
}
\usage{
get_string_ids(identifiers = NULL, echo_query = 0, limit = 5,
  species = 9606, caller_identity)
}
\arguments{
\item{identifiers}{A \code{character} string.}

\item{echo_query}{A \code{logical} in the form '0' (default) or '1'.}

\item{limit}{A \code{numeric}.}

\item{species}{A \code{numeric}.}

\item{caller_identity}{A \code{character} string.}
}
\value{
A \code{tibble}.
\describe{
  \item{queryItem}{(OPTIONAL) your input protein}
  \item{queryIndex}{position of the protein in your input (starting from
  position 0)}
  \item{stringId}{STRING identifier}
  \item{ncbiTaxonId}{NCBI taxon identifier}
  \item{taxonName}{species name}
  \item{preferredName}{common protein name}
  \item{annotation}{protein annotation}
}
}
\description{
Maps common protein names, synonyms and UniProt identifiers into STRING
identifiers.
}
\examples{
\dontrun{
# make a get_string_ids request
get_string_ids(identifiers = c('p53', 'dcdk2'),
               species = 9606)
}

}
\seealso{
\code{\link{network}}
\code{\link{interaction_partners}}
\code{\link{homology}}
\code{\link{homology_best}}
\code{\link{enrichment}}
\code{\link{functional_annotation}}
\code{\link{ppi_enrichment}}

Other API methods: \code{\link{enrichment}},
  \code{\link{functional_annotation}},
  \code{\link{homology_best}}, \code{\link{homology}},
  \code{\link{interaction_partners}},
  \code{\link{network}}, \code{\link{ppi_enrichment}}
}
\concept{API methods}
