% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{homology}
\alias{homology}
\title{Retrieving similarity scores of the protein set}
\source{
https://string-db.org/cgi/help.pl?subpage=api%23retrieving-similarity-scores-of-the-protein-set
}
\usage{
homology(identifiers = NULL, species = 9606, caller_identity)
}
\arguments{
\item{identifiers}{A \code{character} string.}

\item{species}{A \code{numeric}.}

\item{caller_identity}{A \code{character} string.}
}
\value{
A \code{tibble}.
\describe{
  \item{stringId_A}{STRING identifier (protein A)}
  \item{ncbiTaxonId_A}{NCBI taxon identifier (protein A)}
  \item{stringId_B}{STRING identifier (protein B)}
  \item{ncbiTaxonId_B}{NCBI taxon identifier (protein B)}
  \item{bitscore}{Smith-Waterman alignment bit score}
}
}
\description{
Retrieve the protein similarity scores between the input proteins.
}
\examples{
\dontrun{
# make a homology request
homology(identifiers = c('TP53', 'CDK2'))
}

}
\seealso{
\code{\link{get_string_ids}}
\code{\link{network}}
\code{\link{interaction_partners}}
\code{\link{homology_best}}
\code{\link{enrichment}}
\code{\link{functional_annotation}}
\code{\link{ppi_enrichment}}

Other API methods: \code{\link{enrichment}},
  \code{\link{functional_annotation}},
  \code{\link{get_string_ids}},
  \code{\link{homology_best}},
  \code{\link{interaction_partners}},
  \code{\link{network}}, \code{\link{ppi_enrichment}}
}
\concept{API methods}
