% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{homology_best}
\alias{homology_best}
\title{Retrieving best similarity hits between species}
\source{
https://string-db.org/cgi/help.pl?subpage=api%23retrieving-best-similarity-hits-between-species
}
\usage{
homology_best(identifiers = NULL, species = 9606, species_b,
  caller_identity)
}
\arguments{
\item{identifiers}{A \code{character} string.}

\item{species}{A \code{numeric}.}

\item{species_b}{A \code{numeric}.}

\item{caller_identity}{A \code{character} string.}
}
\value{
A \code{tibble}.
\describe{
  \item{stringId_A}{STRING identifier (protein A)}
  \item{ncbiTaxonId_A}{NCBI taxon identifier (protein A)}
  \item{stringId_B}{STRING identifier (protein B)}
  \item{ncbiTaxonId_B}{NCBI taxon identifier (protein B)}
  \item{bitscore}{Smith-Waterman alignment bit score}
}
}
\description{
Retrieve the best (highest) protein similarity hit between different species
in your input.
}
\examples{
\dontrun{
# make a homology_best request
homology_best(identifiers = 'CDK1',
              species_b = 10090)
}

}
\seealso{
\code{\link{get_string_ids}}
\code{\link{network}}
\code{\link{interaction_partners}}
\code{\link{homology}}
\code{\link{enrichment}}
\code{\link{functional_annotation}}
\code{\link{ppi_enrichment}}

Other API methods: \code{\link{enrichment}},
  \code{\link{functional_annotation}},
  \code{\link{get_string_ids}}, \code{\link{homology}},
  \code{\link{interaction_partners}},
  \code{\link{network}}, \code{\link{ppi_enrichment}}
}
\concept{API methods}
