% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{interaction_partners}
\alias{interaction_partners}
\title{Getting all the STRING interaction partners of the protein set}
\source{
https://string-db.org/cgi/help.pl?subpage=api%23getting-all-the-string-interaction-partners-of-the-protein-set
}
\usage{
interaction_partners(identifiers = NULL, limit = 5, species = 9606,
  required_score, caller_identity)
}
\arguments{
\item{identifiers}{A \code{character} string.}

\item{limit}{A \code{numeric}.}

\item{species}{A \code{numeric}.}

\item{required_score}{A \code{numeric}. Values are from 0 to 1000.}

\item{caller_identity}{A \code{character} string.}
}
\value{
A \code{tibble}.
\describe{
  \item{stringId_A}{STRING identifier (protein A)}
  \item{stringId_B}{STRING identifier (protein B)}
  \item{preferredName_A}{common protein name (protein A)}
  \item{preferredName_B}{common protein name (protein B)}
  \item{ncbiTaxonId}{NCBI taxon identifier}
  \item{score}{combined score}
  \item{nscore}{gene neighborhood score}
  \item{fscore}{gene fusion score}
  \item{pscore}{phylogenetic profile score}
  \item{ascore}{coexpression score}
  \item{escore}{experimental score}
  \item{dscore}{database score}
  \item{tscore}{textmining score}
}
}
\description{
Gets all the STRING interaction partners of your proteins.
}
\examples{
\dontrun{
# make a interaction_partners request
interaction_partners(identifiers = c('TP53', 'CDK2'),
                     limit = 10)
}

}
\seealso{
\code{\link{get_string_ids}}
\code{\link{network}}
\code{\link{homology}}
\code{\link{homology_best}}
\code{\link{enrichment}}
\code{\link{functional_annotation}}
\code{\link{ppi_enrichment}}

Other API methods: \code{\link{enrichment}},
  \code{\link{functional_annotation}},
  \code{\link{get_string_ids}},
  \code{\link{homology_best}}, \code{\link{homology}},
  \code{\link{network}}, \code{\link{ppi_enrichment}}
}
\concept{API methods}
