% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_url}
\alias{make_url}
\title{Make a URL}
\usage{
make_url(database = "string-db.org", access = "api", format = "tsv",
  request = "resolve", parameters = NULL)
}
\arguments{
\item{database}{A \code{character}. Possible values are 'string-db.org'
(default) or 'string.embl.de' for the STRING database or 'stitch.embl.de'
for the STITCH sister database.}

\item{access}{A \code{character}. Possible values are 'api' (default) or
services (not used).}

\item{format}{A \code{character}. Only 'tsv' is used.}

\item{request}{A \code{character}. Possibel values are 'resolve' (default),
'abstracts', 'actions', 'interactors' or 'interactions'. A corresponding
request with the name 'requstList' is also available when querying for more
than one item.}

\item{parameters}{A \code{list}. Different requests require different
parameters. The allowed parameters are documented in individual request
functions.}
}
\value{
A \code{character} string of the URL.
}
\description{
Make a valid URL for the STRING API request. The functions translates STRING
required URL fields into \code{modify_url} arguments.
}
\examples{
# make a resolve request for ADD
make_url(parameters = list(identifier = 'ADD'))

}
