% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{network}
\alias{network}
\title{Getting the STRING network interactions}
\source{
https://string-db.org/cgi/help.pl?subpage=api%23getting-the-string-network-interactions
}
\usage{
network(identifiers = NULL, species = 9606, required_score, add_nodes,
  caller_identity)
}
\arguments{
\item{identifiers}{A \code{character} string.}

\item{species}{A \code{numeric}.}

\item{required_score}{A \code{numeric}. Values are from 0 to 1000.}

\item{add_nodes}{A \code{numeric}.}

\item{caller_identity}{A \code{character} string.}
}
\value{
A \code{tibble}.
\describe{
  \item{stringId_A}{STRING identifier (protein A)}
  \item{stringId_B}{STRING identifier (protein B)}
  \item{preferredName_A}{common protein name (protein A)}
  \item{preferredName_B}{common protein name (protein B)}
  \item{ncbiTaxonId}{NCBI taxon identifier}
  \item{score}{combined score}
  \item{nscore}{gene neighborhood score}
  \item{fscore}{gene fusion score}
  \item{pscore}{phylogenetic profile score}
  \item{ascore}{coexpression score}
  \item{escore}{experimental score}
  \item{dscore}{database score}
  \item{tscore}{textmining score}
}
}
\description{
Retrieves the network interactions for your input protein(s) in various
text based formats.
}
\examples{
\dontrun{
# make a network request
network(identifiers = c('TP53', 'EGFR', 'CDK2'),
           required_score = 400)
}

}
\seealso{
\code{\link{get_string_ids}}
\code{\link{interaction_partners}}
\code{\link{homology}}
\code{\link{homology_best}}
\code{\link{enrichment}}
\code{\link{functional_annotation}}
\code{\link{ppi_enrichment}}

Other API methods: \code{\link{enrichment}},
  \code{\link{functional_annotation}},
  \code{\link{get_string_ids}},
  \code{\link{homology_best}}, \code{\link{homology}},
  \code{\link{interaction_partners}},
  \code{\link{ppi_enrichment}}
}
\concept{API methods}
