% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{ppi_enrichment}
\alias{ppi_enrichment}
\title{Getting protein-protein interaction enrichment}
\source{
https://string-db.org/cgi/help.pl?subpage=api%23getting-protein-protein-interaction-enrichment
}
\usage{
ppi_enrichment(identifiers = NULL, species = 9606, required_score,
  caller_identity)
}
\arguments{
\item{identifiers}{A \code{character} string.}

\item{species}{A \code{numeric}.}

\item{required_score}{A \code{numeric}. Values are from 0 to 1000.}

\item{caller_identity}{A \code{character} string.}
}
\value{
A \code{tibble}.
\describe{
 \item{number_of_nodes}{number of proteins in your network}
 \item{number_of_edges}{number of edges in your network}
 \item{average_node_degree}{mean degree of the node in your network}
 \item{local_clustering_coefficient}{average local clustering coefficient}
 \item{expected_number_of_edges}{expected number of edges based on the nodes
  degrees}
 \item{p_value}{significance of your network having more interactions than expected}
}
}
\description{
Tests if your network has more interactions than expected.
}
\examples{
\dontrun{
# make a ppi_enrichment request
ppi_enrichment(identifiers = c('trpA','trpB','trpC','trpE','trpGD'))
}

}
\seealso{
\code{\link{get_string_ids}}
\code{\link{network}}
\code{\link{interaction_partners}}
\code{\link{homology}}
\code{\link{homology_best}}
\code{\link{enrichment}}
\code{\link{functional_annotation}}

Other API methods: \code{\link{enrichment}},
  \code{\link{functional_annotation}},
  \code{\link{get_string_ids}},
  \code{\link{homology_best}}, \code{\link{homology}},
  \code{\link{interaction_partners}}, \code{\link{network}}
}
\concept{API methods}
