% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringapi.R
\docType{data}
\name{query_map_string}
\alias{query_map_string}
\title{Query Map for STRING API (v11)}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26 rows and 7 columns.}
\usage{
query_map_string
}
\description{
The object documents the different types of requests and their query
paramaters that are allowed by the API.
}
\keyword{datasets}
