% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_content}
\alias{format_content}
\title{Format the content of response.}
\usage{
format_content(resp, format = "tsv")
}
\arguments{
\item{resp}{An object of class \code{response}.}

\item{format}{A \code{character}. Only 'tsv' is used.}
}
\value{
A \code{tibble}.
}
\description{
Capture the API response in text format then reformat into tibble.
}
\examples{
# load example response object
fl <- system.file('extdata', 'resp.rda', package = 'stringapi')
load(fl)

# format content
format_content(resp)

}
